// src/store/interfaces/actionTypes.ts
export enum ActionTypes {
  SAVE_ANIME_LIST = 'SAVE_ANIME_LIST',
  ADD_TO_COLLECTION = 'ADD_TO_COLLECTION',
  UPDATE_TO_COLLECTION = 'UPDATE_TO_COLLECTION',
  REMOVE_FROM_COLLECTION = 'REMOVE_FROM_COLLECTION',
  ADD_COLLECTION = 'ADD_COLLECTION',
  UPDATE_COLLECTION = 'UPDATE_COLLECTION',
  REMOVE_COLLECTION = 'REMOVE_COLLECTION',
  SAVE_PAGE_INFO = 'SAVE_PAGE_INFO',
}